Initial import of the JFPatch as a service (RO Build service).
This has been developed over the last few days from the bare Pyromaniac
build environment, and should now contain a base on which the service
can actually be created. The JFPatch is just a snapshot. The Pyromaniac
is the 'last docker version' so isn't ideal in its management, but this
does mean less duplication.

In the future we'll pull in the different parts of the environment
from the artifact server.