Add a result collection class; rework all the building into a class.
The results are now in a single class, which is used with callbacks
to accumulate the output so that we can write it out to a remote
server if necessary. These functions are called by the docker
and pyronativeserver objects.

The main code that was just in the body of the build.py is now in
a class, and can be called in stages - this allows additional code
to be inserted between the parts, such as to add debug or to
debug issues.