Add execution timeout; support for BASTXT files; tidier JSON and binary.
An execution timeout of 60 seconds is imposed by the 'timeout' tool.
This should ensure that we never run away with code being left running.

Basic Text files are now included in the buildables, which means that
you can submit a simple BASIC file and get out the text that it generated
but not any binary, as there are no clipboard operations (unless the
user adds them).

JSON output now comes out pretty printed if the data is small.
The Binary output now works!