Add option for debug to CLI; clean up better on failures.
The CLI tool can now specify a comma-separated list of debug flags
for the pyromaniac system. This makes it easier to see what's going
on when needed.

When we call the 'close' method on the builder, the rosource will
clean up the temporary directory it created. This should make it
less messy when something goes wrong.