Fix the zip file extraction support.
Zip files which contain amu files are now built properly. The files
were not being processed properly with the Acorn extensions not being
extracted as expected. The dates were not extracted from the files
which meant that the objects were always stamped with the same
timestamp. We now stamp them with the original timestamp from the
zip file, and then for any files that are 'buildables' we restamp
them with the current timestamp so that they are newer than any
of the existing files.