Add simple makefile parser.
The makefile parser is intended to make it possible to determine
the targets available, and the directories and objects that need
to be present to build. It's not an especially complex parser -
merely one that follows many of the rules for the makefile to
allow the extraction of this information.

In theory it could produce the necessary commands to build the
targets, but that isn't its goal. It's not even clear if the
information it provides is actually that useful in building the
targets.