Add support for finding targets and commands for a Makefile.
The commands can be determined for the Makefile contents. This means
that it's possible to check the content for what directories should
be present, and maybe create them in advance.

Finding linkable targets is less likely to be useful at the moment,
but might be used to offer the option to suggest targets to build.