Update ROBuild to use the directory creation details from the Makefile.
When the build is being performed with a Makefile, we can determine
what would be built. Those components will then indicate what directories
need to be present - and we can create them as we go. This is useful for
those cases where the directories aren't included in the zip archive that
was supplied.

We still don't know what the output targets are, but that's a separate
issue.