Add support for the streamed output from the build.
Streamed output means that we get results as they happen through a
callback. The results are streamed on to the same thread that did
the request. Output lines are accumulated through a heartbeat, to
ensure that we output data regularly and we don't supply single
character callbacks for each letter (because that's how we tend to
write things on RISC OS).