Tidy up the websocket client and server, and JSON server for live use.
The live site is currently pointing at the test system, and with that,
the configuration has been updated so that the tools and the commands
that access the system are able to work that way.

The JSON server is now on a port alongside the WebSocket service.
The responses no longer the values as lists for most of the values
returned. This might be changed in the future, but it's simplest to
just leave them as bare strings in most cases.