Add 'basic' HTML client that can communicate with the back end.
The 'basic' HTML client is actually a quite powerful WebSocket client
which can pass the supplied source to the back end, display the messages
from the build process, handle throwback and clipboard data and allow
the user to download the output file afterward, if any.

It's surprising that it works quite so well!