Fix the character set used by the frontend; add selection of wss:.
The front end's character set wasn't selecting UTF-8, so the build status
was coming out in an ugly way. It's now set with a meta directive.

The use of ws: when connected by HTTPS is insecure and rejected by the
browser. So I've made it select ws: or wss: depending on whether the
HTTP or HTTPS was used.