Checkpoint for the JFPatch file format documentation.
The JFPatch documentation is partially converted to HTML. There's a
lot missing, and it could do with some examples littering through
it, but it seems to be at least vaguely working as it stands.

This change only really covers the Format file; it doesn't cover the
!Summary, which might contain more information to be included.
Similarly the !Help itself may include information about the historic
nature of the application that may fit into the about pages at some
point.