Add missing documentation for some module features, macros.
The modules did not define a number of fields in the documentation.
Many of the filters, the MessageFile, Resources and ImageFS/FS were
not documented at all. These have now been added. Similarly the
documentation has been updated to add in the Macros which were not
documented at all (and are surprisingly non-obvious).

Some of the styling has been tidied up a little, and the SVG for
file_xxx has been cleaned up so that it has the correct header.