Add codemirror and JFPatch assembler mode.
The JFPatch assembler mode provides all the colouring for the
tool, as it currently stands. I believe that most of the colouring
is sane. It might not be the colours that everyone likes, but that's
not necessarily the point. It's got the right bits for what we need -
and includes ARM and BBC BASIC colouring within the sections that
of the file as appropriate.