Add the source code editor to the front end.
The front end is now able to show you the source code and let you
edit it and then send it to the server for building. It's likely
that the workflow isn't quite right for a real development
environment, but... honestly, who cares, it's a freakin' build
tool for RISC OS in a browser.