Add the additional context colouring for the source code.
We now have context colouring for C files, Pascal files, and Perl files
in addition to the JFPatch files. We have the ability to select the
Basic Text and ObjAsm files for colouring, but currently I've not written
the modes to do that colouring.

The colouring source is now built conditionally based on whether we
have enabled the different supported modes. Strictly that's only for the
front end. The back end is always able to support them.

The preprocessor I've used is one that I wrote for Doom, Heretic and
Hexen, with additional features for the processing of the conditionals,
which it turned out I hadn't implemented. I might as well have written
a clean pre-processor rather than reusing that one, but hey-ho.