Add ARM assembler mode to the BBC BASIC colouring.
The ARM assembler colouring is complicated because it is embedded in
the BASIC, and actually provides a lot of the same functions that the
ARM assembler uses, but has to do it within the constraints of having
line numbers (or not), and also handling BBC BASIC style assembler
(OPT, and FN inline) and expressions that are BBC BASIC style (which
are not coloured as such right now).

The BBC BASIC examples also updated with inline assembler as well.