Add help icon and guidance for how to use the service.
The service is relatively simple, to me, but others may not have much
of a clue what the buttons do or the workflow. It's not as polished
(by far) as Compiler Explorer, so people may need a bit more help.
In the future, it might be nice to simplify the interface so that it
doesn't follow the back end service interface, but right now it is
what it is and we can help make it easier to use.

The help box toggles with the button, but also hides itself when you
press one of the create or upload buttons.

This change also renames the 'Save' button to 'Send' because that better
represents what it does, and I don't really want to confuse people with
thinking that it's going to save to some storage.