Create a structure diagram for how the service works.
I love diagrams. Pictures, they say, are worth a thousand words, and
I don't like writing words, so pictures are cool. On the other hand,
it's a lot easier to write words than to get Graphviz to do what you
want.

Diagram shows the parts of the system, with the containers they live
in. It can optionally show the JSON server, as well as the WS server,
but it makes the diagram less clear, so I've turned the option off
right now.