Added the structure diagram to the About page.
The structure diagram is now included in the About page, and the
technologies used is expanded upon a little bit. It's kinda amusing
that I've spent multiple days on getting the syntax colouring right,
but haven't put the time into making JFPatch create 32bit modules
yet. 'cos that bit is boring.