Front end fixes to handle WebSocket disconnect better.
When we're disconnected this appears as a message in the output
window, which helps to understand what's going on beyond the status
that appears beside the help icon. The states where we've disconnected
and have source which needs to be re-sent are now recognised and messages
added to the output window. In the future this workflow should probably
be updated to be more efficient - initially it was intended that we only
transfer data when the user actually requests it, keeping them in
control, but it seems that that's not a very useful flow.

It'd still be possible to repeatedly transfer lots of data if the
connection kept going up and down at the moment but the user would have
to actively send the data. Probably for the sizes we're talking about
this is not especially relevant, and we could just take the hit of
sending the data on reconnect.