Add 'load example' button.
We now have a button to load files, in addition to the create new
document and the uplaod document. That might be a confusing collection,
so maybe they should be rationalised. However, for now this works, and
I've not got time to be playing around with it.

We take the examples from the local submodule, so the state of the
repository when the service was deployed is used. This could be
reorganised in the future.