Fix for failure to actually invoke the tool on zip archives.
When a zip archive is supplied we determine what the tool we should
use from the filetypes of the files, and then we report the command
line to use and execute it. Only, at some point, the list of buildable
files in the zip archive changed from a tuple of name and filetype
to a RISCOSName object that translates unix names to RISCOS names
and filetypes (and vice-versa). But it was only changed in the single
file code path.

This change ensures that we use the correct names, and introduces the
example which performs the patching. You don't get an editor window
when you've got a zip file (I had considered multiple editors, one
for each file, but... complexity).