Improve rendering of the images; add a large icon.
The images weren't rendering well - the imagemagick conversion appears
to be affected by whether a viewbox specification is given in the SVG,
even though that viewbox is not used. These are now included, and this
should make the icons appear better.

There is now a large patched icon generated. This was intended to be
used for the repository card, but it's too much of a faff to get it
generating in the right shape, so I've given up for now.