Start reworking the build to take from Artifactory.
Rather than being hand crafted, and building on the demo Pyromaniac
docker container, this pulls some of its resources directly from
Artifactory. This means that we should have versions of the components
which are automatically build from the CI, and being fed into the
build service when it is updated.

However, at present, there's still a lot of components that need to
be embedded directly.