Fix for output accumulator not writing out properly on heartbeat.
The 'heartbeat' callback is meant to be invisible to the calling user,
and causes output to flush to the stream if no further output to a
newline has been seen in a period (usually a few seconds). The heartbeat
was unfortunately being exposed as a 'heartbeat' action, with the
output attached. That wasn't the intention - it should be written
out as a plain 'output' action.

Updated the docker container name to match the container that we
build.