Replace the hardcoded RO Zip info parser with rozipinfo.
The rozipinfo module is a lot more complex and contains a lot more
than we currently need for the reading of the Zip archives, but we
should be able to use it for the creation of Zip archives for the
outputs in the future.

Also fixed some zip file decoding problems when the Makefile does
not supply any targets (in this case because it uses the includes
to include the targets it expects to build).