Initial support for the .robuild.yml build configuration.
If a build configuration file (.robuild.yml) is found, we'll process
it and try to build using the definition given within the file. The
file format is currently in flux, as I'm playing with it, but seemed
to work with the LineEditor build - but without artifacts. They will
need to be added separately.