Update front end to recognise Zip archives, with a new icon.
The front end can now recognise a Zip archive being produced as the
output, and has an SVG icon for the zip file, so that it looks nice.
We have special case code that can turn the &a91 types into .zip
with the correct media type. Although the files produced are in
RISC OS format (ie, not NFS encoding), I would still expect the use
to be for RISC OS, without any special handling, so using the
RISC OS extra information is what users would expect, I think.