Fix for multiple dependency rules not being appended properly.
When multiple dependency rules were present, as used by older versions
of amu, the rules would fail to be appended properly - instead of the
list accumulating the subsequent dependencies, it would nest them.

This caused exceptions because the system was not expecting that form
of list - it should be a flat set of dependencies.