Make the runtime consistent between interfaces; add ping endpoint.
The maximum execution time has been pushed to 10 minutes, and made
consistent between the JSON and WS interfaces.

There's an endpoint for 'ping' now, which just returns OK for the JSON
service. This allows it to be used as a healthcheck, which is needed for
load balancer use within the AWS ALB. The ALB is needed because the
prior solution of using CloudFront was failing due to the maximum timeout
of 60 seconds.