Disable the use of bold in the ANSIText output.
The bold text means that the output when rendered in the CI systems
looks a bit odd - with it coming out brighter than the text surrounding
it. Because the use of colours really isn't all that significant in
the build system, I've disabled the bold usage.

This means that we will treat all the standard colours as if they have
maximum brightness for the purposes of colour matching, and will never
use the bold attribute. 24bit and 8bit colour sequences can still be
generated, as they have not been disabled.