Update the RISC OS source processor to not touch buildables (optionally).
An internal option to not build the buildables files from the archive,
has been added. The intention of touching those files was to force the
rebuilding of the sources if the user uploads object files. Obviously
if they sent the files to the build service, they want them to be built.
But that might be wrong in the future, so there's an internal option
to disable that behaviour. It's not exposed anywhere yet, but it's at
least available for testing.