Fix for P2C header builds, permissions on Lib$Dir; add Alias command.
The P2C header file wasn't in the right place, so couldn't be found
by the build.

Lib$Dir wasn't writable so libraries couldn't be written there, which
was required by the riscos-build-online tool.

The *Alias command is now supported, to make it easier to mimic a
RISC OS Classic system that has been booted.