Add support for websocket options.
WebSocket connections can request options, and read the current
options for the session. There are two features that I want to
expose in this way - one is a global timeout, which is currently
hard coded in the server, and the other is the ability to disable
ansitext VDU implementation, reverting to just the plain text output,
without any control codes.

The latter is not yet implemented, but the former is present in
this change.