Update JFPatch tool to pull from Artifactory.
The JFPatch tool source is now pulled from the build version of
canonical source, rather than being embedded in the service source.
This should mean that we can update it independant of the service -
keeping the two concerns quite separate!

The new JFPatch that's coming in should support 32bit module builds
as well, which is also handy.