Update docker container to use the latest Pyromaniac build.
The docker container is now based on the latest (0.17.1900)
pyromaniac-build container. This means that we no longer need to
collect the resources ourselves (except for JFPatch). The new container
will, however, contain JFPatch which is pulled from the resources.
This most recent version of JFPatch supports the 32bit module builds.

The change in structure has necessitated some changes to the service,
as the docker invocation is now slightly different - we now use the
configuration files more heavily, and these are better layered.