Harmonise the stylesheet and code with the resource site.
The resource site provides a lot of information using the same basic
style and libraries that were used in the JFPatch-as-a-service site.
However, they have been extended and the style improved in some cases.
We bring those changes back to this site, and add some more descriptions
in some of the pages.