Add initial documentation of the CI usage.
The beginnings of some documentation about how you can use the
service in CI. At the moment this describes how you use the JSON
API and the WS API through robuild-client within plain shell
commands. There are placeholders for the GitHub and GitLab
configuration, which will be documented later.

Shell colouring has been introduced so that we have a nicer to
look at presentation of the script lines.