Finish updating the CI documentation to give a guide on use.
The CI documentation describes how you might perform the build with
the JSON API and the WebSockets API through robuild-client. The
examples are given in unix shell.

CI wrappers are explained for the GitHub and GitLab CI systems.
Jenkins use would depend on the type of job that was used, and
the manner in which the author wishes to use it; it's easier to
leave that to them - the shell scripting examples will be easy
to transplant.