Update service to be deployable through Ansible.
The service can now be deployed through Ansible, with configuration
to pull the necessary docker container on service start, and with
configurable overrides. This ensures that the system can theoretically
keep up with the deployed versions from my docker repository. In reality,
it probably won't need to be, because it's not expected to auto-scale.