Rework makefile builds to be able to return linkables.
Linkables are now able to be returned as artifacts in the automatically
processed Makefiles in archives. There were some problems with the
handling of the makefiles and we never actually returned the linkables
that it had determined were present anyhow. This was work that had been
deferred - it's now complete.

The examples have been expanded to include C application and C module
code. The documentation has been expanded to give more explanation of
the API and additional links.