Add download links for JFShared and JFPatch.
The JFPatch source has been updated so that the full application
can now be built. It's really terrible, and honestly, the service
is much better in general (and it's not great really!), but we may
as well have a working binary release. I'm genuinely not sure how
well it runs on RISC OS Classic, but I'm also not that fussed.