Upgraded docker image to speed up builds; added config option to cli.py.
The build jobs were running very slowly. This was traced to the
tracepoint support, which was triggered in OS_SynchroniseCodeAreas,
which caused each call to take significantly longer. This has been made
more efficient, and is now ~23,000x faster.

During debugging it was useful to add the extra configuration information
to trigger the timings statistics. There is now a --config option to make
it easier to test.