Update the site branding to build.riscos.online style.
The site is now switching over the build.riscos.online style, from
the old JFPatch-as-a-service branding. The build tool is pretty much
the same, but this is acknowledging that it's a generic build service.

The menu items have also moved around slightly to de-emphasise the
JFPatch side of the service.