Update the configuration to allow runtime configuration of input.
The input settings are relatively benign and can allow us to test some
operations which would otherwise cause the system to terminate whilst
processing user input. We may as well allow this, for the
input.eof_effect setting.

The EOF default is now to cause a Reset, which should cause the system
to exit.