Changed the help menu's labels to be more helpful.
Coming off of the results of my survey, the help menu has been updated
to hopefully make it easier to find what people may want.

The Help menu was referring to CI, and the RISC OS Build configuration,
which wasn't very helpful. They now say explicitly 'Automation with git'
and '.robuild.yaml format' to make it very clear what they talk about.
And the documentation pages are prefixed by Docs to help that.