Make the JSON response include error messages instead of Badness.
The JSON response was invalid if we had an error - it would return
a plain text message 'Badness: <exception message>'. That's not very
helpful.

This change introduces a JSON response when we have a server error.
We still report as a 500 - as that means server error - but the body
is now a well formed JSON structure.