Fix mistakes in the CI examples page.
The CI examples page has been updated with a couple of small fixes
for the shell code that wasn't coming out right - a couple of path
mistakes meant that it might not work if used as-is.

The links to different places have been updated as well, to include
the CI templates repo, and the simple BASIC file building repo.